Fix async drain integration test
According to the debug logs this test introduced the ginkgo node
timeouts in CI.

GinkgoRecover is needed to collect failures from a go routine.

After adding "defer GinkgoRecover" to the drain test, it would sometimes
hang indefinitely. The "Eventually" was added to make sure the async
test exits.

Also removed the "lifecycle" test. The lifecycle part had already been
removed two years ago and it was now a duplicate of tests in deploy_test.go.
